Make sure that the remote origin url is correct

+++ git_force_update.csh
@@ -1,5 +1,5 @@
 #!/bin/tcsh -f
-# Time-stamp: <2015-10-21 13:55:51 kmodi>
+# Time-stamp: <2016-02-23 11:04:12 kmodi>
 
 # Usage: source git_force_update.csh <YOUR .emacs.d PATH>
 # Example: source git_force_update.csh ~/.emacs.d
@@ -43,13 +43,25 @@ if ( ! -d ${emacs_config_dir} ) then
     echo "Cloning .emacs.d to ${emacs_config_dir} .."
     git clone http://www.github.com/kaushalmodi/.emacs.d ${emacs_config_dir}
 else
-    if ( -d ${emacs_autobkp_dir} ) then
-        \rm -rf ${emacs_autobkp_dir} # Delete old backup if present
+    cd ${emacs_config_dir}
+    set git_remote_origin_url = `git config --get remote.origin.url`
+    set git_remote_origin_url_ = `echo ${git_remote_origin_url} | sed 's/.*\(github\.com.*\.emacs\.d\).*/\1/'`
+    # Ensure that the we are not overriding a different git managed dir
+    if ( ${git_remote_origin_url_} == "github.com/kaushalmodi/.emacs.d" ) then
+        if ( -d ${emacs_autobkp_dir} ) then
+            \rm -rf ${emacs_autobkp_dir} # Delete old backup if present
+        endif
+        echo "Backing up existing ${emacs_config_dir} to ${emacs_autobkp_dir} .."
+        \cp -rf ${emacs_config_dir} ${emacs_autobkp_dir}
+        echo "Force updating .emacs.d to ${emacs_config_dir} .."
+        gfu
+    else
+        echo "ERROR: The remote URL for this git repo is '${git_remote_origin_url}'."
+        echo "Please try again using a different directory argument for downloading this .emacs.d."
+        exit 0
     endif
-    echo "Backing up existing ${emacs_config_dir} to ${emacs_autobkp_dir} .."
-    \cp -rf ${emacs_config_dir} ${emacs_autobkp_dir}
-    echo "Force updating .emacs.d to ${emacs_config_dir} .."
-    cd ${emacs_config_dir}; gfu
+    unset {git_remote_origin_url}
+    unset {git_remote_origin_url_}
 endif
 echo ''
